refactor: move mocked trpc to utils directory
