+++ src/stories/utils/mockTrpc.ts
@@ -1,8 +1,8 @@
-import { TRPCError } from '@trpc/server'
+import { type TRPCError } from '@trpc/server'
 import { getHTTPStatusCodeFromError } from '@trpc/server/http'
 import { TRPC_ERROR_CODES_BY_KEY } from '@trpc/server/rpc'
 import { createTRPCMsw } from 'msw-trpc'
-import { AppRouter } from '~/server/modules/_app'
+import { type AppRouter } from '~/server/modules/_app'
 import { getBaseUrl } from '~/utils/getBaseUrl'
 
 export const trpcMsw = createTRPCMsw<AppRouter>({
